refactor(media): extract scene router
