+++ src/routers/media.ts
@@ -4,19 +4,12 @@ import path from "path";
 
 import { getConfig } from "../config";
 import Image from "../types/image";
-import Scene from "../types/scene";
 import { logger } from "../utils/logger";
+import sceneRouter from "./scene";
 
 const router = Router();
 
-router.get("/scene/:scene", async (req, res, next) => {
-  const scene = await Scene.getById(req.params.scene);
-
-  if (scene && scene.path) {
-    const resolved = path.resolve(scene.path);
-    res.sendFile(resolved);
-  } else next(404);
-});
+router.use("/scene", sceneRouter);
 
 router.get("/image/path/:path", async (req, res) => {
   const pathParam = (req.query as Record<string, string>).path;
