+++ src/routers/scene.ts
@@ -1,0 +1,17 @@
+import { Router } from "express";
+import path from "path";
+
+import Scene from "../types/scene";
+
+const router = Router();
+
+router.get("/:scene", async (req, res, next) => {
+  const scene = await Scene.getById(req.params.scene);
+
+  if (scene && scene.path) {
+    const resolved = path.resolve(scene.path);
+    res.sendFile(resolved);
+  } else next(404);
+});
+
+export default router;
